default type changed to double from int

+++ app/src/main/java/myFirstProject/App.java
@@ -30,7 +30,7 @@ public class App {
         java.util.ArrayList<Number> inputList = new java.util.ArrayList<>();
         while (sc1.hasNext())
         {
-            int value = Integer.parseInt(sc1.next().replaceAll("\\s",""));
+            double value = Double.parseDouble(sc1.next().replaceAll("\\s",""));
             inputList.add(value);
         }
         sc1.close();
